cmake, reorganised directories, PING PONG bugfix
