--- config
@@ -1,3 +1,0 @@
-SERVER_ADDR irc.freenode.net
-SERVER_PORT 8001
-BOT_NICK botche
