+++ src/connection.h
